Provided one global DeviceConnector instance. Moved typedefs from utils to a new file.
