+++ tests/infra/utils.py
@@ -2,44 +2,13 @@
 #
 # SPDX-License-Identifier: Apache-2.0
 
-from dataclasses import dataclass
-from enum import Enum
-from typing import Any, Callable, Mapping, Optional, Sequence, Union
-
 import jax
 import jax.numpy as jnp
-from flax import linen, nnx
 from jax import export
 
-
-@dataclass
-class Workload:
-    executable: Callable
-    args: Sequence[Any]
-    kwargs: Optional[Mapping[str, Any]] = None
-
-    def __post_init__(self):
-        # If kwargs is None, initialize it to an empty dictionary.
-        if self.kwargs is None:
-            self.kwargs = {}
-
-    def execute(self) -> Any:
-        return self.executable(*self.args, **self.kwargs)
-
-
-class Framework(Enum):
-    JAX = "jax"
-    TORCH = "torch"
-    NUMPY = "numpy"
-
-
-# Convenience alias. Could be used to represent jax.Array, torch.Tensor, np.ndarray, etc.
-Tensor = Union[jax.Array]
-
-# Convenience alias. Could be used to represent nnx.Module, torch.nn.Module, etc.
-# NOTE nnx.Module is the newest API, linen.Module is legacy but it is used in some
-# huggingface models.
-Model = Union[nnx.Module, linen.Module]
+from .device_runner import run_on_cpu
+from .types import Framework, Tensor
+from .workload import Workload
 
 
 def _str_to_dtype(dtype_str: str, framework: Framework = Framework.JAX):
@@ -50,6 +19,7 @@ def _str_to_dtype(dtype_str: str, framework: Framework = Framework.JAX):
         raise ValueError(f"Unsupported framework: {framework.value}.")
 
 
+@run_on_cpu
 def random_tensor(
     shape: tuple,
     dtype: str = "float32",
